made GPIO tables explicit, add ESP32-C3 targets, add DCServo HAL Device

+++ src/HAL_JSON/ScriptEngine/Parser/HAL_JSON_SCRIPT_ENGINE_Parser_Expressions.cpp
@@ -502,7 +502,7 @@ namespace HAL_JSON {
                             anyError = true;
                         }
                     } else {
-                        HALValue halBracketSubscriptValue(0); // just need a dummy value
+                        HALValue halBracketSubscriptValue((uint32_t)0); // just need a dummy value
                         HALValue halValue;
                         readResult = device->read(halBracketSubscriptValue, halValue);
                         if (readResult != HALOperationResult::Success) {
@@ -519,7 +519,7 @@ namespace HAL_JSON {
                             anyError = true;
                         }
                     } else {
-                        HALValue halBracketSubscriptValue(0); // just need a dummy value
+                        HALValue halBracketSubscriptValue((uint32_t)0); // just need a dummy value
                         HALValue halValue(HALValue::Type::TEST); // unset type
                         writeResult = device->write(halBracketSubscriptValue, halValue);
                         if (writeResult != HALOperationResult::Success) {
